requirements.txt
# ══════════════════════════════════════════════════════════
#  Fraud Detection System — Dependencies
# ══════════════════════════════════════════════════════════

# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Gradient Boosting Models
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Visualization (training / notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# API & Serving
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Config
pyyaml>=6.0.0

# Optional — Deep Learning (uncomment if using LSTM)
# tensorflow>=2.15.0
# keras>=3.0.0

# Optional — Experiment tracking
# mlflow>=2.9.0
# shap>=0.43.0

# Optional — Development
# jupyterlab>=4.0.0
# pytest>=7.4.0
# black>=23.11.0
